backend/requirements.txt
# Advanced Search Engine Backend - Requirements

# FastAPI and Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis for Caching
redis==5.0.1
aioredis==2.0.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Search and NLP
elasticsearch==8.11.0
whoosh==2.7.4
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
transformers==4.35.2
sentence-transformers==2.2.2

# Web Scraping and Crawling
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests-html==0.10.0

# Data Processing and Analysis
pandas==2.1.4
numpy==1.25.2
networkx==3.2.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Rate Limiting and Security
slowapi==0.1.9
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Background Tasks
celery==5.3.4
kombu==5.3.4
